Add files via upload
Here is what I have for the project ideas that we came up with. Sorry the hammer isn't on here, I just stopped at 4 since that is what the homework description says. 

We still need to do the project rubric. I would suggest doing in such a way as to emphasize us choosing based on group interest and feasibility.